Refactored common code into lib/, using symlinks where appropriate
